Decided to use an array to store input values
Using an array with primitive values is faster than an ArrayList.  There
is also no needed conversion when working with the values.  A problem
with this is that the default value for each index is 0.0 because it is
an array of doubles.  How will we deal with all of the extra zeros in
the array?  We will have to find a way to find the last value in the
array, for instance if the last input in the array is 0.0 the program
needs to know that 0.0 is the end of the list and not a nonzero number.
(the default size of the array is 100 to allow for a reasonable amount
of inputs from the user)